+++ EngineLayer/SFIni.cpp
@@ -110,12 +110,12 @@ DWORD SFIni::GetPathName(LPTSTR lpBuffer, DWORD dwBufSize) const
 	if (lpBuffer != NULL)
 	{
 		_tcsncpy(lpBuffer, m_pszPathName, dwBufSize);
-		dwLen = _tcslen(lpBuffer);
+		dwLen = (DWORD)_tcslen(lpBuffer);
 	}
 	else
 	{
 		// just calculate the required buffer size
-		dwLen = _tcslen(m_pszPathName);
+		dwLen = (DWORD)_tcslen(m_pszPathName);
 	}
 	return dwLen;
 }
@@ -139,7 +139,7 @@ DWORD SFIni::GetString(LPCTSTR lpSection, LPCTSTR lpKey, LPTSTR lpBuffer, DWORD 
 		*lpBuffer = _T('\0');
 
 	LPTSTR psz = __GetStringDynamic(lpSection, lpKey, lpDefault);
-	DWORD dwLen = _tcslen(psz);
+	DWORD dwLen = (DWORD)_tcslen(psz);
 
 	if (lpBuffer != NULL)
 	{
@@ -208,7 +208,7 @@ DWORD SFIni::GetArray(LPCTSTR lpSection, LPCTSTR lpKey, LPTSTR lpBuffer, DWORD d
 		if (lpBuffer == NULL)
 		{
 			// just calculate the required buffer size
-			const DWORD MAX_LEN = _tcslen(psz) + 2;
+			const DWORD MAX_LEN = (DWORD)(_tcslen(psz) + 2);
 			LPTSTR p = new TCHAR[MAX_LEN + 1];
 			dwCopied = __StringSplit(psz, p, MAX_LEN, lpDelimiter, bTrimString);
 			delete [] p;
@@ -386,7 +386,7 @@ BOOL SFIni::WriteChar(LPCTSTR lpSection, LPCTSTR lpKey, TCHAR c) const
 DWORD SFIni::GetDataBlock(LPCTSTR lpSection, LPCTSTR lpKey, LPVOID lpBuffer, DWORD dwBufSize, DWORD dwOffset) const
 {
 	LPTSTR psz = __GetStringDynamic(lpSection, lpKey);
-	DWORD dwLen = _tcslen(psz) / 2;
+	DWORD dwLen = (DWORD)(_tcslen(psz) / 2);
 	if (dwLen <= dwOffset)
 	{
 		delete [] psz;
@@ -802,11 +802,11 @@ DWORD SFIni::__StringSplit(LPCTSTR lpString, LPTSTR lpBuffer, DWORD dwBufSize, L
 	if (lpDelimiter != NULL && *lpDelimiter == _T('\0'))
 	{
 		_tcsncpy(lpBuffer, lpString, dwBufSize - 1);
-		return _tcslen(lpBuffer);
+		return (DWORD)_tcslen(lpBuffer);
 	}
 
 	LPTSTR pszDel = (lpDelimiter == NULL) ? _tcsdup(DEF_PROFILE_DELIMITER) : _tcsdup(lpDelimiter);
-	const DWORD DEL_LEN = _tcslen(pszDel);
+	const DWORD DEL_LEN = (DWORD)_tcslen(pszDel);
 	LPTSTR lpTarget = lpBuffer;
 
 	// Search through lpString for delimiter matches, and extract sub strings out
@@ -819,7 +819,7 @@ DWORD SFIni::__StringSplit(LPCTSTR lpString, LPTSTR lpBuffer, DWORD dwBufSize, L
 		if (bTrimString)
 			__TrimString(pszSeg);
 
-		const DWORD SEG_LEN = _tcslen(pszSeg);
+		const DWORD SEG_LEN = (DWORD)_tcslen(pszSeg);
 		const DWORD COPY_LEN = min(SEG_LEN, dwBufSize - dwCopied);
 
 		// Need to avoid buffer overflow
@@ -841,7 +841,7 @@ DWORD SFIni::__StringSplit(LPCTSTR lpString, LPTSTR lpBuffer, DWORD dwBufSize, L
 	if (bTrimString)
 		__TrimString(pszSeg);
 
-	const DWORD SEG_LEN = _tcslen(pszSeg);
+	const DWORD SEG_LEN = (DWORD)_tcslen(pszSeg);
 	const DWORD COPY_LEN = min(SEG_LEN, dwBufSize - dwCopied);
 
 	if (COPY_LEN > 0)
@@ -865,7 +865,7 @@ BOOL SFIni::ParseDNTString(LPCTSTR lpString, SUBSTRPROC lpFnStrProc, LPVOID lpPa
 		return FALSE;
 
 	LPCTSTR p = lpString;
-	DWORD dwLen = _tcslen(p);
+	DWORD dwLen = (DWORD)_tcslen(p);
 
 	while (dwLen > 0)
 	{
@@ -873,7 +873,7 @@ BOOL SFIni::ParseDNTString(LPCTSTR lpString, SUBSTRPROC lpFnStrProc, LPVOID lpPa
 			return FALSE;
 
 		p = &p[dwLen + 1];
-		dwLen = _tcslen(p);
+		dwLen = (DWORD)_tcslen(p);
 	}
 	return TRUE;
 }
@@ -911,7 +911,7 @@ BOOL CALLBACK SFIni:: __KeyPairProc(LPCTSTR lpString, LPVOID lpParam)
 
 	// trim
 	__TrimString(psz);
-	DWORD dwNameLen = _tcslen(psz);
+	DWORD dwNameLen = (DWORD)_tcslen(psz);
 	DWORD dwCopyLen = 0;
 	
 	//copy to the buffer
@@ -1040,7 +1040,7 @@ BOOL SFIni::__TrimString(LPTSTR lpString)
 		return FALSE;
 
 	BOOL bTrimmed = FALSE;
-	int nLen = _tcslen(lpString);
+	int nLen = (int)_tcslen(lpString);
 
 	// '\n' and '\r' are actually not possible in this case, but anyway...
